Refactor code to use blueprints
Moved routing code to portfolio.py
Replaced Flask Navigation package with native Flask code
Adjusted URLs to start from "/"
Added trailing slash to routes and to url test
Fixed typo in 404 template
