+++ app/templates/404.html
@@ -2,36 +2,37 @@
 <html lang="en">
 
 <head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <meta property="og:title" content="Personal Portfolio">
-    <meta property="og:description" content="My Personal Portfolio">
-    <meta property="og:url" content="{{ url }}">
+  <!-- Required meta tags -->
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+  <meta property="og:title" content="Personal Portfolio">
+  <meta property="og:description" content="My Personal Portfolio">
+  <meta property="og:url" content="{{ url }}">
 
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-    <link
-        href="https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap"
-        rel="stylesheet">
+  <link rel="preconnect" href="https://fonts.gstatic.com">
+  <link
+    href="https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap"
+    rel="stylesheet">
 
-    <link lang='sass' rel="stylesheet" href="./static/styles/404.css">
-    <link rel='icon' href='./static/img/favicon.ico' type='image/x-icon' />
-    <title>{{ title }}</title>
+  <link lang='sass' rel="stylesheet" href="/static/styles/404.css">
+  <link rel='icon' href='/static/img/favicon.ico' type='image/x-icon' />
+  <title>{{ title }}</title>
 </head>
 
 <body>
-    <div class="split left">
-  <div class="centered">
-    <img src='./static/img/miyabi_cry.jpg' alt="Miyabi Crying">
+  <div class="split left">
+    <div class="centered">
+      <img src='/static/img/miyabi_cry.jpg' alt="Miyabi Crying">
+    </div>
   </div>
-</div>
 
-<div class="split right">
-  <div class="centered">
-    <h1>404</h1>
-    <p>Well...this is embarrising</p>
-    <a href={{ url_for("index") }}>Go Home</a>
+  <div class="split right">
+    <div class="centered">
+      <h1>404</h1>
+      <p>Well...this is embarrassing.</p>
+      <a href={{ url_for("portfolio.index") }}>Go Home</a>
+    </div>
   </div>
-</div>
 </body>
+
 </html>
