+++ app/templates/contact.html
@@ -1,4 +1,5 @@
 {% extends 'base.html' %}
+{% set active_page = "contact" %}
 
 {% block content %}
 <div>
