+++ app/templates/index.html
@@ -1,4 +1,5 @@
 {% extends 'base.html' %}
+{% set active_page = "index" %}
 
 {% block content %}
 <div>
