+++ app/templates/projects.html
@@ -1,4 +1,5 @@
 {% extends 'base.html' %}
+{% set active_page = "projects" %}
 
 {% block content %}
 <div>
