+++ app/templates/resume.html
@@ -1,4 +1,5 @@
 {% extends 'base.html' %}
+{% set active_page = "resume" %}
 
 {% block content %}
 <div>
