+++ tests/app_test.py
@@ -13,16 +13,16 @@ def test_index(client):
     assert response.status_code == 200
 
 
-def test_index(client):
-    response = client.get("/projects", content_type="html/text")
+def test_projects(client):
+    response = client.get("/projects/", content_type="html/text")
     assert response.status_code == 200
 
 
-def test_index(client):
-    response = client.get("/contact", content_type="html/text")
+def test_contact(client):
+    response = client.get("/contact/", content_type="html/text")
     assert response.status_code == 200
 
 
-def test_index(client):
-    response = client.get("/resume", content_type="html/text")
+def test_resume(client):
+    response = client.get("/resume/", content_type="html/text")
     assert response.status_code == 200
